FindSubWords is a Golang Coding Challenge
Given a list of words, like word.list, the FindSubWords.go program's
compiled executable will find the longest compound-word in the list.
These compounds are a concatenation of other sub-words that exist in
the list. The program allows the user to input different data via the
commmand-line by specifying a file name. Any programming language can
be used, but Go is preferred.